requirements.txt
Django==2.2.7
psycopg2==2.8.4
Pillow==6.2.1
djangorestframework==3.10.3
django-filter==2.2.0


# celery==4.3.0
# amqp==2.5.1
#gunicorn==19.9.0

# pytz==2019.03 #auto